ISIS-2727: moves InteractionLayer and InteractionServie from runtime to applib
